+++ TestRoot/normalizeStackTraces.lua
@@ -9,14 +9,7 @@ local function serialize(val: string | { [any]: any }, config, indentation, dept
 		val = removeRootFromStackTrace(val)
 	end
 
-	return printer(
-		val,
-		config,
-		indentation,
-		depth,
-		refs,
-		printer
-	)
+	return printer(val, config, indentation, depth, refs, printer)
 end
 
 local function containsStackTrace(val: string): boolean
@@ -25,7 +18,8 @@ local function containsStackTrace(val: string): boolean
 end
 
 local function test(val: any): boolean
-	return (typeof(val) == "string" and containsStackTrace(val)) or (typeof(val) == "table" and typeof(val.message) == "string" and containsStackTrace(val.message))
+	return (typeof(val) == "string" and containsStackTrace(val))
+		or (typeof(val) == "table" and typeof(val.message) == "string" and containsStackTrace(val.message))
 end
 
 return {